Move `AbstractImage`, `FriendImage`, `GroupImage` from mirai-core-api to mirai-core, close #1670
